Calendar Agent/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-dotenv>=1.0
anthropic>=0.40
google-auth>=2.35
google-auth-oauthlib>=1.2
google-auth-httplib2>=0.2
google-api-python-client>=2.149
itsdangerous>=2.2
pydantic>=2.10
python-dateutil>=2.9
pytz>=2024.2
